Making a BigM mix-in class and moving multiple bigm onto the base class
